requirements.txt
nose
sphinx
pylru
enum
numpy
pillow
protobuf
tinys3
